+++ src/app/components/products/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 import { CartService, ProductsService } from '../../../shared/services';
 import { ProductModel } from 'src/app/models';
@@ -10,8 +11,9 @@ import { ProductModel } from 'src/app/models';
   templateUrl: './product-list.component.html',
   styleUrls: ['./product-list.component.scss']
 })
-export class ProductListComponent implements OnInit {
-  products$: Observable<ProductModel[]>;
+export class ProductListComponent implements OnInit, OnDestroy {
+  destroy$ = new Subject<void>();
+  products: Promise<ProductModel[]>;
 
   constructor(
     private router: Router,
@@ -20,11 +22,29 @@ export class ProductListComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
-    this.products$ = this.productsService.getProducts();
+    this.products = this.productsService.getProducts();
   }
 
-  onBuy(id: number): void {
-    this.cartService.addProduct(id);
+  ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.complete();
+  }
+
+  onBuy(product: ProductModel): void {
+    const observer = {
+      error: (err: any) => console.log(err)
+    };
+    const cartProduct = this.cartService.getCartProductById(product.id);
+    if (cartProduct) {
+      const increaseQuantityProduct = this.cartService.getIncreaseQuantityProduct(cartProduct);
+      this.cartService.updateProducts(increaseQuantityProduct)
+        .pipe(takeUntil(this.destroy$))
+        .subscribe(observer);
+    } else {
+      this.cartService.addProduct(product)
+        .pipe(takeUntil(this.destroy$))
+        .subscribe(observer);
+    }
   }
 
   onViewProduct(product: ProductModel): void {