+++ src/app/models/app-settings.model.ts
@@ -1,0 +1,3 @@
+export interface AppConfigModel {
+  readonly darkTheme: boolean;
+}