+++ src/app/shared/services/index.ts
@@ -7,3 +7,4 @@ export * from './generator.factory';
 export * from './constants.service';
 export * from './auth.service';
 export * from './dialog.service';
+export * from './app-settings.service';