+++ src/app/shared/services/local-storage.service.ts
@@ -3,7 +3,8 @@ import { Injectable, InjectionToken } from '@angular/core';
 export enum LocalStorageKeys {
   id = 'id',
   name = 'name',
-  cart = 'cart'
+  cart = 'cart',
+  appSettings = 'app-settings'
 }
 
 export const STORAGE = new InjectionToken<LocalStorageService>('Storage');
@@ -13,15 +14,15 @@ export const STORAGE = new InjectionToken<LocalStorageService>('Storage');
 })
 export class LocalStorageService {
   set(key: LocalStorageKeys, value: string): void {
-    localStorage.setItem(LocalStorageKeys[key], value);
+    localStorage.setItem(key, value);
   }
 
   get(key: LocalStorageKeys): string {
-    return localStorage.getItem(LocalStorageKeys[key]);
+    return localStorage.getItem(key);
   }
 
   remove(key: LocalStorageKeys): void {
-    localStorage.removeItem(LocalStorageKeys[key]);
+    localStorage.removeItem(key);
   }
 
   clear(): void{