+++ src/app/shared/services/products.service.ts
@@ -1,34 +1,78 @@
 import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 
-import { productsMock } from '../../mocks/products.mock';
 import { ProductModel } from '../../models';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductsService {
-  getProducts(): Observable<ProductModel[]> {
-    return new Observable(observe => {
-      setTimeout(() => {
-        observe.next(productsMock);
-      }, 300);
-    });
+  private productsUrl = 'http://localhost:3000/products';
+
+  constructor(private http: HttpClient) {}
+
+  getProducts(): Promise<ProductModel[]> {
+    return this.http
+      .get(this.productsUrl)
+      .toPromise()
+      .then(response => response as ProductModel[])
+      .catch(this.handleError);
   }
 
-  getProductById(id: number): Observable<ProductModel> {
-    return of(productsMock.find(item => item.id === id));
+  getProductById(id: number): Promise<ProductModel> {
+    const url = `${this.productsUrl}/${id}`;
+
+    return this.http
+      .get(url)
+      .toPromise()
+      .then(response => response as ProductModel)
+      .catch(this.handleError);
   }
 
-  createProduct(product: ProductModel): void {
-    productsMock.push(product);
+  createProduct(product: ProductModel): Promise<ProductModel> {
+    const url = this.productsUrl;
+    const body = JSON.stringify(product);
+    const options = {
+      headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+    };
+
+    return this.http
+      .post(url, body, options)
+      .toPromise()
+      .then(response => response as ProductModel)
+      .catch(this.handleError);
   }
 
-  updateProduct(product: ProductModel): void {
-    const i = productsMock.findIndex(t => t.id === product.id);
 
-    if (i > -1) {
-      productsMock.splice(i, 1, product);
-    }
+  updateProduct(product: ProductModel): Promise<ProductModel>  {
+    const url = `${this.productsUrl}/${product.id}`;
+    const body = JSON.stringify(product);
+    const options = {
+        headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+      };
+
+    return this.http
+      .put(url, body, options)
+      .toPromise()
+      .then(response => response as ProductModel)
+      .catch(this.handleError);
+  }
+
+  deleteTask(product: ProductModel): Promise<ProductModel> {
+    const url = `${this.productsUrl}/${product.id}`;
+
+    return (
+      this.http
+        .delete(url)
+        .toPromise()
+        // json-server return empty object
+        // so we don't use .then(...)
+        .catch(this.handleError)
+    );
+  }
+
+  private handleError(error: any): Promise<any> {
+    console.error('An error occurred', error);
+    return Promise.reject(error.message || error);
   }
 }